first split
