+++ projects/core/src/data-providers/rest-data-provider.ts
@@ -8,7 +8,7 @@ import type {
 } from '../data-interfaces.js'
 
 import { UrlBuilder } from '../../urlBuilder.js'
-import { buildRestDataProvider, retry } from '../buildRestDataProvider.js'
+import { buildRestDataProvider } from '../buildRestDataProvider.js'
 import type { ApiClient } from '../context.js'
 import { customUrlToken, Filter } from '../filter/filter-interfaces.js'
 import type { EntityMetadata, FindOptions } from '../remult3/remult3.js'
@@ -19,7 +19,7 @@ export class RestDataProvider implements DataProvider {
   constructor(
     private apiProvider: () => ApiClient,
     private entityRequested?: (entity: EntityMetadata) => void,
-  ) {}
+  ) { }
   public getEntityDataProvider(entity: EntityMetadata): RestEntityDataProvider {
     this.entityRequested?.(entity)
     return new RestEntityDataProvider(
@@ -126,13 +126,12 @@ export function findOptionsFromJson(
 }
 
 export class RestEntityDataProvider
-  implements EntityDataProvider, ProxyEntityDataProvider
-{
+  implements EntityDataProvider, ProxyEntityDataProvider {
   constructor(
     private url: () => string,
     private http: () => RestDataProviderHttpProvider,
     private entity: EntityMetadata,
-  ) {}
+  ) { }
   query(
     options: EntityDataProviderFindOptions,
     aggregateOptions: EntityDataProviderGroupByOptions,
@@ -297,7 +296,7 @@ export class RestEntityDataProvider
     return this.http()
       .put(
         this.url() +
-          (id != '' ? '/' + encodeURIComponent(id) : '?__action=emptyId'),
+        (id != '' ? '/' + encodeURIComponent(id) : '?__action=emptyId'),
         this.toJsonOfIncludedKeys(data),
       )
       .then((y) => this.translateFromJson(y))
@@ -336,99 +335,6 @@ export class RestEntityDataProvider
         data.map((data) => this.translateToJson(data)),
       )
       .then((y) => y.map((y: any) => this.translateFromJson(y)))
-  }
-}
-
-export class RestDataProviderHttpProviderUsingFetch
-  implements RestDataProviderHttpProvider
-{
-  constructor(
-    private fetch?: (
-      input: RequestInfo,
-      init?: RequestInit,
-    ) => Promise<Response>,
-  ) {}
-  async get(url: string) {
-    return await retry(async () =>
-      this.myFetch(url).then((r) => {
-        return r
-      }),
-    )
-  }
-  put(url: string, data: any) {
-    return this.myFetch(url, {
-      method: 'put',
-      body: JSON.stringify(data),
-    })
-  }
-  delete(url: string) {
-    return this.myFetch(url, { method: 'delete' })
-  }
-  async post(url: string, data: any) {
-    return await retry(() =>
-      this.myFetch(url, {
-        method: 'post',
-        body: JSON.stringify(data),
-      }),
-    )
-  }
-
-  myFetch(
-    url: string,
-    options?: {
-      method?: string
-      body?: string
-    },
-  ): Promise<any> {
-    const headers: any = {}
-    if (options?.body) headers['Content-type'] = 'application/json'
-    if (
-      typeof window !== 'undefined' &&
-      typeof window.document !== 'undefined' &&
-      typeof (window.document.cookie !== 'undefined')
-    )
-      for (const cookie of window.document.cookie.split(';')) {
-        if (cookie.trim().startsWith('XSRF-TOKEN=')) {
-          headers['X-XSRF-TOKEN'] = cookie.split('=')[1]
-        }
-      }
-    return (this.fetch || fetch)(url, {
-      credentials: 'include',
-      method: options?.method,
-      body: options?.body,
-      headers,
-    })
-      .then((response) => {
-        return onSuccess(response)
-      })
-      .catch(async (error) => {
-        let r = await error
-        throw r
-      })
-  }
-}
-
-function onSuccess(response: Response) {
-  if (response.status == 204) return
-  if (response.status >= 200 && response.status < 300) return response.json()
-  else {
-    throw response
-      .json()
-      .then((x) => {
-        return {
-          ...x,
-          message: x.message || response.statusText,
-          url: response.url,
-          status: response.status,
-        }
-      })
-      .catch(() => {
-        throw {
-          message: response.statusText,
-          url: response.url,
-          status: response.status,
-        }
-      })
   }
 }
